+++ src/Seeders/Seeder.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Codemonster\Database\Seeders;
+
+abstract class Seeder
+{
+    /**
+     * Run the seeder.
+     */
+    abstract public function run(): void;
+}